bgweb struct conversion and better engine interface functions

+++ server/types/game.go
@@ -14,11 +14,11 @@ func NewDices() [2]int {
 }
 
 type Turn struct {
-	ID     int64      `json:"id"`
-	GameId int64      `json:"game-id"`
-	User   string     `json:"user"`
-	Time   time.Time  `json:"time"`
-	Moves  [][2]uint8 `json:"moves"`
+	ID     int64     `json:"id"`
+	GameId int64     `json:"game-id"`
+	User   string    `json:"user"`
+	Time   time.Time `json:"time"`
+	Moves  [][2]int  `json:"moves"`
 }
 
 type Game struct {
@@ -32,8 +32,8 @@ type Game struct {
 	End    time.Time `json:"end"`
 	Status string    `json:"status"`
 
-	P1Checkers [25]uint8 `json:"p1checkers"` // arr[0] is bar
-	P2Checkers [25]uint8 `json:"p2checkers"` // arr[0] is bar
+	P1Checkers [25]int `json:"p1checkers"` // arr[0] is bar
+	P2Checkers [25]int `json:"p2checkers"` // arr[0] is bar
 
 	Double      uint64 `json:"double"`
 	DoubleOwner string `json:"double-owner"`
